Bob Kurtz reported a problem with the mandelbrot demo, with the following error:
2012-08-15 10:00:05,118 SHOUT [DOTS] - ....Unable to read /org/cougaar/demo/mandelbrot/util/default_palette.map: java.lang.RuntimeException: Unable to find resource
Pallete /org/cougaar/demo/mandelbrot/util/default_palette.map contains 0 colors

The directory structure at the high level is a bit different than last
time; it's set to ".../cougaar-12.7/CougaarApps/tutorials/mandelbrot"
(in my Cougaar 12.6 installation, the structure was simply
".../cougaar-12.6/mandelbrot" but I'm looking to differentiate the
applications from the Cougaar core. In any event, it seems that the path
is relative (but starting from where?), so it doesn't look like it should
matter.  The default_palette.map file does exist in the
src/org/cougaar/demo/mandelbrot/util directory.

Dave Montana provided the solution:
In mandelbrot/build.xml, we were not copying over the *.map file before building the jar.
Therefore, it was working in eclipse but not when running from the zips.